pyproject.toml
[project]
name = "mealplan"
version = "0.1.0"
description = "Personal meal planning optimization with LP/QP"
requires-python = ">=3.9"

dependencies = [
    "numpy>=1.24",
    "scipy>=1.11",
    "pandas>=2.0",
    "typer[all]>=0.9",
    "rich>=13.0",
    "pyyaml>=6.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0",
    "pytest-cov",
    "ruff",
    "mypy",
]

[project.scripts]
mealplan = "mealplan.cli:app"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/mealplan"]

[tool.ruff]
line-length = 100
target-version = "py39"

[tool.ruff.lint]
select = ["E", "F", "I", "UP"]

[tool.mypy]
python_version = "3.9"
strict = true
